docs(observability): formalize metrics conventions, align existing metrics to TagList (#499, #500)
Document metrics naming conventions (file placement, metric naming, tag
conventions, instrument types, ActivitySource) in coding-standards.md and
CLAUDE.md. Move RateLimitingMetrics and BulkheadMetrics to Diagnostics/
folder. Refactor all 3 existing metrics classes from KeyValuePair to
TagList for zero-allocation tag passing. Add Microsoft.Extensions.
Diagnostics.Testing to Directory.Packages.props for MetricCollector<T>
unit test assertions.
