+++ Directory.Packages.props
@@ -44,6 +44,7 @@
     <PackageVersion Include="Microsoft.EntityFrameworkCore.Relational" Version="10.*" />
     <PackageVersion Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="10.*" />
     <PackageVersion Include="Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore" Version="10.*" />
+    <PackageVersion Include="Microsoft.Extensions.Diagnostics.Testing" Version="10.*" />
     <PackageVersion Include="Microsoft.EntityFrameworkCore.SqlServer" Version="10.*" />
     <PackageVersion Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="10.*" />
     <PackageVersion Include="Microsoft.EntityFrameworkCore.InMemory" Version="10.*" />
