+++ src/Granit.Http.Bulkhead/Extensions/BulkheadServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 using Granit.Http.Bulkhead.Abstractions;
+using Granit.Http.Bulkhead.Diagnostics;
 using Granit.Http.Bulkhead.Exceptions;
 using Granit.Http.Bulkhead.Internal;
 using Granit.Http.Bulkhead.Options;
