+++ src/Granit.Http.Bulkhead/TenantPartitionedBulkhead.cs
@@ -1,6 +1,7 @@
 using System.Threading.RateLimiting;
 using Granit.Core.MultiTenancy;
 using Granit.Http.Bulkhead.Abstractions;
+using Granit.Http.Bulkhead.Diagnostics;
 using Granit.Http.Bulkhead.Exceptions;
 using Granit.Http.Bulkhead.Internal;
 using Granit.Http.Bulkhead.Options;
