--- src/Granit.RateLimiting/RateLimitingMetrics.cs
@@ -1,32 +1,0 @@
-using System.Diagnostics.Metrics;
-
-namespace Granit.RateLimiting;
-
-/// <summary>
-/// Metrics for the rate limiting module.
-/// Meter: <c>Granit.RateLimiting</c>.
-/// </summary>
-public sealed class RateLimitingMetrics
-{
-    public const string MeterName = "Granit.RateLimiting";
-
-    private readonly Counter<long> _allowedCounter;
-    private readonly Counter<long> _rejectedCounter;
-
-    public RateLimitingMetrics(IMeterFactory meterFactory)
-    {
-        Meter meter = meterFactory.Create(MeterName);
-        _allowedCounter = meter.CreateCounter<long>("granit.ratelimiting.requests.allowed", description: "Number of requests allowed by rate limiting.");
-        _rejectedCounter = meter.CreateCounter<long>("granit.ratelimiting.requests.rejected", description: "Number of requests rejected by rate limiting.");
-    }
-
-    public void RecordAllowed(string policyName, string? tenantId) =>
-        _allowedCounter.Add(1,
-            new KeyValuePair<string, object?>("policy", policyName),
-            new KeyValuePair<string, object?>("tenant_id", tenantId ?? "global"));
-
-    public void RecordRejected(string policyName, string? tenantId) =>
-        _rejectedCounter.Add(1,
-            new KeyValuePair<string, object?>("policy", policyName),
-            new KeyValuePair<string, object?>("tenant_id", tenantId ?? "global"));
-}
