+++ src/Granit.RateLimiting/TenantPartitionedRateLimiter.cs
@@ -1,5 +1,6 @@
 using Granit.Core.MultiTenancy;
 using Granit.RateLimiting.Abstractions;
+using Granit.RateLimiting.Diagnostics;
 using Granit.RateLimiting.Internal;
 using Granit.RateLimiting.Options;
 using Granit.Security;
