+++ tests/Granit.RateLimiting.Tests/RateLimitMiddlewareTests.cs
@@ -2,6 +2,7 @@ using System.Diagnostics.Metrics;
 using Granit.Core.MultiTenancy;
 using Granit.RateLimiting.Abstractions;
 using Granit.RateLimiting.Attributes;
+using Granit.RateLimiting.Diagnostics;
 using Granit.RateLimiting.Exceptions;
 using Granit.RateLimiting.Internal;
 using Granit.RateLimiting.Options;
