+++ tests/Granit.RateLimiting.Tests/TenantPartitionedRateLimiterTests.cs
@@ -1,6 +1,7 @@
 using System.Diagnostics.Metrics;
 using Granit.Core.MultiTenancy;
 using Granit.RateLimiting.Abstractions;
+using Granit.RateLimiting.Diagnostics;
 using Granit.RateLimiting.Internal;
 using Granit.RateLimiting.Options;
 using Granit.Security;
